requirements.txt
pytest==7.4.4
pytest-mock==3.12.0
pillow==10.2.0
